rename core pkg to common
