+++ cow_py/common/abi/ComposableCow.json
