+++ cow_py/common/constants.py
