--- cow_py/core/__init__.py
@@ -1,4 +1,0 @@
-from cow_py.core import constants, cow_error, abis, chains, config
-from cow_py.core.chains import utils
-
-__all__ = ["constants", "cow_error", "abis", "chains", "config", "utils"]
