+++ cow_py/subgraphs/deployments.py
@@ -1,4 +1,4 @@
-from cow_py.core.chains import Chain
+from cow_py.common.chains import Chain
 from dataclasses import dataclass
 from enum import Enum
