+++ cow_py/web3/provider.py
@@ -1,7 +1,7 @@
 from typing import Dict
 
 import web3
-from cow_py.core.chains import Chain
+from cow_py.common.chains import Chain
 
 DEFAULT_PROVIDER_NETWORK_MAPPING = {
     Chain.MAINNET: "https://eth.llamarpc.com",
