+++ tests/core/test_cow_error.py
@@ -1,4 +1,4 @@
-from cow_py.core.cow_error import (
+from cow_py.common.cow_error import (
     CowError,
 )  # Adjust the import path according to your project structure
